addressed the rest of PR comments so far...
