+++ src/main/java/org/folio/circulation/domain/representations/RequestProperties.java
@@ -5,6 +5,7 @@ public class RequestProperties {
 
   public static final String STATUS = "status";
   public static final String NAME = "name";
+  public static final String REQUEST_TYPE = "requestType";
   public static final String PROXY_USER_ID = "proxyUserId";
   public static final String POSITION = "position";
   public static final String HOLD_SHELF_EXPIRATION_DATE = "holdShelfExpirationDate";
